[lexical-playground] Refactor: Migrate CollapsiblePlugin to CollapsibleExtension
Converts the legacy React CollapsiblePlugin into a declarative
CollapsibleExtension using defineExtension, registering the collapsible
nodes, INSERT_COLLAPSIBLE_COMMAND, node transforms, and keyboard
navigation handlers through the extension system.
