+++ packages/lexical-playground/src/plugins/CollapsibleExtension/CollapsibleContainerNode.ts
