+++ packages/lexical-playground/src/plugins/CollapsibleExtension/CollapsibleContentNode.ts
