+++ packages/lexical-playground/src/plugins/CollapsibleExtension/CollapsibleTitleNode.ts
