+++ packages/lexical-playground/src/plugins/CollapsibleExtension/CollapsibleUtils.ts
